Jud Participates (#1)
# Description of changes
## Enumerated
1. Removed excess command line args and added necessary information to
the .env file. (If we end up actually pursuing this idea we won't need
to be passing command line args. Rather, you will define page
information in the actual code. Likely the spaceID will be directly
embedded into the program.)
2. Code cleaning. I moved a bunch of stuff into packages.
3. Added documentation.
## Pontificated
Here I have done a bit of code cleaning and documentation. I like that
you've messed with this a bit. I think if you are generally on board we
need to hammer down why we are doing this and what we can do to
accomplish our clearly defined objectives. I think this is a great idea
to tinker with. I have provided my desired directives in the
documentation. Feel free to expand/contract what I have provided.

Anywho...